SOURCES CHATBOT/chatbot-CNCAC-main/backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chatbot-cncac-backend"
version = "1.0.0"
description = "ChatDocAI Backend - AI chatbot for French notaries with RAG capabilities"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "ChatDocAI Team", email = "contact@chatdocai.fr"}
]
keywords = ["fastapi", "rag", "neo4j", "openai", "notary"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Legal Industry",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Framework :: FastAPI",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic[email]>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "supabase>=2.3.0",
    "aiofiles>=23.2.1",
    "httpx>=0.26.0",
    "minio>=7.2.0",
    "openai>=1.50.0",
    "neo4j>=5.25.0",
    "tiktoken>=0.10.0",
    "loguru>=0.7.0",
    # Document parsing avec Docling
    "docling>=2.0.0",
    "docling-core>=2.0.0",
    "docling-parse>=2.0.0",
    # Support additionnel pour formats spécifiques
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "odfpy>=1.4.0",
    "extract-msg>=0.48.0",
    "pandas",
    "rdflib",
    "PyMuPDF",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "pytest-cov==6.0.0",
    "pytest-mock==3.14.0",
    "pytest-timeout==2.3.1",
    "httpx==0.28.1",
    "faker==33.1.0",
    "factory-boy==3.3.1",
    "coverage[toml]==7.6.10",
    # Linting & Formatting
    "ruff>=0.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    # Type stubs
    "types-python-dateutil",
    "types-requests",
]

test = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "pytest-cov==6.0.0",
    "pytest-mock==3.14.0",
    "pytest-timeout==2.3.1",
    "httpx==0.28.1",
    "faker==33.1.0",
    "factory-boy==3.3.1",
    "coverage[toml]==7.6.10",
]

[project.urls]
Homepage = "https://github.com/juliendabert/chatbot-CNCAC"
Documentation = "https://chatdocai.fr/docs"
Repository = "https://github.com/juliendabert/chatbot-CNCAC"
Issues = "https://github.com/juliendabert/chatbot-CNCAC/issues"

[project.scripts]
chatdocai = "src.main:main"
chatdocai-ingest = "scripts.ingestion_pipeline:main"
chatdocai-enrich = "src.agents.graph_enrichment_agent:main"

[tool.setuptools]
package-dir = {"" = "."}
packages = ["src", "scripts", "tests"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json", "*.txt"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Tests that take a long time to run",
]
addopts = [
    "--strict-markers",
    "--tb=short",
    "-v",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=xml",
    "--cov-report=html",
    "--cov-fail-under=70",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
    "*/migrations/*",
    "*/venv/*",
    "*/backend/venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "def __str__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@abc.abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"

[tool.ruff]
target-version = "py311"
line-length = 100
fix = true
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.isort]
known-third-party = ["fastapi", "pydantic", "starlette"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | venv
  | backend/venv
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
strict_optional = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true